+++ R/promises.R
@@ -58,7 +58,7 @@ as.promise.mirai <- function(x)
       query <- function()
         if (unresolved(x))
           later::later(query, delay = 0.1) else
-            if (is_error_value(value <- .subset2(x, "value")))
+            if (is_error_value(value <- .subset2(x, "data")))
               reject(value) else
                 resolve(value)
       query()