+++ NEWS.md
@@ -5,6 +5,7 @@
   + '.args' will now coerece to a list if an object other than a list is supplied, rather than error.
   + '.signal' argument removed - now all 'mirai' signal if daemons are set up.
 * `everywhere()` now returns invisible NULL in the case the specified compute profile is not set up, rather than error.
+* Improved error messages and handling for daemon/dispatcher connection errors.
 * Requires nanonext >= [0.10.4.9001].
 
 # mirai 0.11.2